% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{add_indices}
\alias{add_indices}
\title{Calculate indices and raster stack}
\usage{
add_indices(raster, file_name, count, indices_df)
}
\arguments{
\item{raster}{a terra SpatRaster object}

\item{file_name}{output file path for saving image file, should not include file extension}

\item{count}{character to append on the end of the file name in the case of multiple images}

\item{indices_df}{data frame object containing list of indices that match user-specified search criteria and band names}
}
\value{
ouput file path containing image file
}
\description{
Calculate indices and raster stack
}
