% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{add_scl_layer}
\alias{add_scl_layer}
\title{Add scene classification layer}
\usage{
add_scl_layer(keep_mask = FALSE, composite, layers)
}
\arguments{
\item{keep_mask}{logical representing if scene classification layer should be returned in STAC query}

\item{composite}{character representing compositing method to apply in STAC query}

\item{layers}{an rsi band mapping object}
}
\value{
an rsi band mapping object
}
\description{
Add scene classification layer
}
