% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{check_bbox_area}
\alias{check_bbox_area}
\title{Check area of interest size}
\usage{
check_bbox_area(sf_object, max_area)
}
\arguments{
\item{sf_object}{an sf object}

\item{max_area}{a numeric value representing maximum area in m^2 to throw a warning message}
}
\description{
Check area of interest size
}
