% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{check_date_in_range}
\alias{check_date_in_range}
\title{Check year of interest within date range}
\usage{
check_date_in_range(start_dt, end_dt, target_yr)
}
\arguments{
\item{start_dt}{character representing start date in format "YYYY-MM-DD"}

\item{end_dt}{character representing end date in format "YYYY-MM-DD"}

\item{target_yr}{character representing a year in format "YYYY"}
}
\value{
character vector representing the target search range
}
\description{
Check year of interest within date range
}
