% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{extract_vals_df}
\alias{extract_vals_df}
\title{Extract values from raster based on sf geometry}
\usage{
extract_vals_df(
  raster,
  buffer_layer,
  col_ID = NULL,
  file_name = NULL,
  add_dates = FALSE
)
}
\arguments{
\item{raster}{a terra SpatRaster object.}

\item{buffer_layer}{an sf object}

\item{col_ID}{a field from the sf object representing a unique ID}

\item{file_name}{a named character vector that stores the file path and the date of acquisition for a .tif file.}

\item{add_dates}{logical value representing if date of acquisition should be appended to column names of output dataframe}
}
\value{
data frame object containing raster values.
}
\description{
Extract values from raster based on sf geometry
}
