% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{find_first_date}
\alias{find_first_date}
\title{Find date of first image within a date range}
\usage{
find_first_date(
  start_date,
  end_date,
  stac_search_fxn,
  aoi_search,
  band_mapping,
  layers_search
)
}
\arguments{
\item{start_date}{character representing start date in format "YYYY-MM-DD"}

\item{end_date}{character representing end date in format "YYYY-MM-DD"}

\item{stac_search_fxn}{one of the \code{get_stac_data} functions from the rsi package}

\item{aoi_search}{an sf object}

\item{band_mapping}{an rsi band mapping object}

\item{layers_search}{either an rsi band mapping object or character vector specifying image assets to keep}
}
\value{
the first date within the date range that an image is captured, if successful
}
\description{
Find date of first image within a date range
}
