% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{indices_df}
\alias{indices_df}
\title{Generate data frame of indices to calculate}
\usage{
indices_df(band_names, index_application = NULL, index_names = NULL)
}
\arguments{
\item{band_names}{character of length > 0. Band names as required by the Awesome Spectral Indices (ASI) repository.}

\item{index_application}{character vector specifying application domains to include in calculations. this value
is based on the application_domain column from the ASI repository.}

\item{index_names}{character vector specifying index names to calculate. This value is based on the
short_name column from the ASI repository.}
}
\value{
data frame object containing list of indices that match user-specified search criteria and band names
}
\description{
Generate data frame of indices to calculate
}
