% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{point_buffer}
\alias{point_buffer}
\title{Add buffer to sf layer}
\usage{
point_buffer(boundary_layer, radius = NULL, max_dist)
}
\arguments{
\item{boundary_layer}{object of class sf. in this package, sf object with point geometry}

\item{radius}{a numeric value representing the distance (in meters) to apply a buffer}

\item{max_dist}{a numeric value representing the maximum distance which will throw a warning}
}
\value{
object of class sf with polygon geometry
}
\description{
Add buffer to sf layer
}
\examples{
point_buffer(sf_point, 100, 1000)
point_buffer(sf_point, 501, 500)
}
