% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{prepare_aoi}
\alias{prepare_aoi}
\title{Prepare area of interest for STAC query}
\usage{
prepare_aoi(aoi_layer, radius = NULL)
}
\arguments{
\item{aoi_layer}{an sf object}

\item{radius}{a numeric value representing the distance (in meters) to apply a buffer}
}
\value{
an sf object
}
\description{
Prepare area of interest for STAC query
}
