% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{process_image_data}
\alias{process_image_data}
\title{Generic image processing function}
\usage{
process_image_data(
  stac_search_fxn,
  aoi,
  date_range,
  band_mapping,
  search_layers = band_mapping,
  mask_name = NULL,
  mask_function = NULL,
  composite = "median",
  resample = "bilinear",
  max_date_range = NULL,
  date_interval = NULL
)
}
\arguments{
\item{stac_search_fxn}{one of the \code{get_stac_data} functions from rsi package}

\item{aoi}{an sf object}

\item{date_range}{a character vector containing a validated date range}

\item{band_mapping}{an rsi band mapping object}

\item{search_layers}{an rsi band mapping object}

\item{mask_name}{the name of the image layer to use for masking}

\item{mask_function}{the function to use to apply mask to image}

\item{composite}{character representing image compositing method}

\item{resample}{character representing image resampling method}

\item{max_date_range}{numeric representing maximum number of days within date range,
used to return a warning message}

\item{date_interval}{interval at which images from a sensor are captured, for example,
Sentinel-2 images are captured every 5 days.}
}
\value{
a named character vector where each element is a file name containing output .tif file
and each element name is either the date of image acquisition (when composite is NULL) or the date
range of image acquisition (when composite is not NULL)
}
\description{
Generic image processing function
}
