% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{s1_dB}
\alias{s1_dB}
\title{Apply backscatter to dB calulcation}
\usage{
s1_dB(raster, file_pth = NULL, counter = NULL)
}
\arguments{
\item{raster}{a terra SpatRaster object}

\item{file_pth}{output file path to save .tif image to}

\item{counter}{character to append on the end of the file name in the case of multiple images}
}
\value{
ouput file path containing image file
}
\description{
Apply backscatter to dB calulcation
}
