% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_facing.R
\name{s2_process}
\alias{s2_process}
\title{Sentinel-2 L2A Processing}
\usage{
s2_process(
  aoi_layer,
  radius = NULL,
  start_dt,
  end_dt,
  uniqueID = NULL,
  layers_sel = NULL,
  composite_method = c("median", "mean", "sum", "min", "max", NULL),
  resample_method = c("bilinear", "average", "rms", "nearest", "gauss", "cube",
    "cubicspline", "lanczos", "average_magphase", "mode"),
  keep_SCL = FALSE,
  apply_mask = TRUE,
  app_domains = NULL,
  idx_names = NULL,
  file_path
)
}
\arguments{
\item{aoi_layer}{}

\item{radius}{}

\item{start_dt}{}

\item{end_dt}{}

\item{uniqueID}{}

\item{layers_sel}{}

\item{composite_method}{}

\item{resample_method}{}

\item{keep_SCL}{}

\item{apply_mask}{}

\item{app_domains}{}

\item{idx_names}{}

\item{file_path}{}
}
\description{
Sentinel-2 L2A Processing
}
