% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{s2_scale}
\alias{s2_scale}
\title{Apply Sentinel-2 L2A surface reflectance conversion}
\usage{
s2_scale(raster, file_pth = NULL, counter = NULL)
}
\arguments{
\item{raster}{a terra SpatRaster object}

\item{file_pth}{character representing file path to save output file to, should not include file extension}

\item{counter}{character to append on the end of the file name in the case of multiple images}
}
\value{
character representing the output file path
}
\description{
Apply Sentinel-2 L2A surface reflectance conversion
}
\examples{
s2_scale(raster_data, file_pth="exfolder/exfile")
}
