% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{select_layers}
\alias{select_layers}
\title{Select layers to return from STAC query}
\usage{
select_layers(layer_list = NULL, band_mapping)
}
\arguments{
\item{layer_list}{character vector representing layers to return in STAC query}

\item{band_mapping}{an rsi band mapping object}
}
\value{
an rsi band mapping object
}
\description{
Select layers to return from STAC query
}
