% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{valid_col_ID}
\alias{valid_col_ID}
\title{Validate sf column ID}
\usage{
valid_col_ID(aoi, column_ID = NULL)
}
\arguments{
\item{aoi}{an sf object}

\item{column_ID}{the field from the sf object to use as a unique ID}
}
\value{
character representing unique column ID from sf object.
}
\description{
Validate sf column ID
}
