% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{valid_dates}
\alias{valid_dates}
\title{Validate date range}
\usage{
valid_dates(start_date, end_date)
}
\arguments{
\item{start_date}{character representing start date in format "YYYY-MM-DD"}

\item{end_date}{character representing end date in format "YYYY-MM-DD"}
}
\value{
character vector of length 2 containing start date and end date to be used in stac query
}
\description{
Validate date range
}
\examples{
valid_dates("2020-01-01","2020-12-31")
}
