% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{validate_method}
\alias{validate_method}
\title{Generic method validation}
\usage{
validate_method(method, valid_methods, param_name)
}
\arguments{
\item{method}{character representing user-defined method}

\item{valid_methods}{character vector representing valid methods}

\item{param_name}{the name of the parameter to be used in output error message}
}
\description{
Generic method validation
}
